fix: resolve all compilation errors after Kotlin migration
- Fix STTCallback/TtsCallback/CommandExecutor interface nullability mismatches
- Replace setCallback() calls with property assignment (callback =)
- Fix NavigationManager Long-to-Float type conversion
- Fix JSONObject HashMap type cast in BaiduSpeechManager
- Fix smart cast issues with local variables
- Fix PoiSearch.Query.isCityLimit property access
- Fix AppDatabase.getInstance() → Room.databaseBuilder()
- Fix SuggestionAdapter lambda → object expression
- Fix packageInfo.signatures null assertion
- Build compiles successfully with only deprecation warnings

+++ app/src/main/java/com/example/voicenavigation/navigation/NavigationManager.kt
@@ -244,8 +244,8 @@ class NavigationManager(private val context: Context) : RouteSearch.OnRouteSearc
 
             currentRouteResult = result
             currentWalkPath = result.paths[0]
-            totalDistance = currentWalkPath!!.distance
-            totalDuration = currentWalkPath!!.duration
+            totalDistance = currentWalkPath!!.distance.toFloat()
+            totalDuration = currentWalkPath!!.duration.toFloat()
             remainingDistance = totalDistance
 
             val steps: List<WalkStep> = currentWalkPath!!.steps
